# Optimize rust code of loop and levenshtein