--- a/loops/rust/code.rs
+++ b/loops/rust/code.rs
@@ -9,11 +9,11 @@ fn main() {
     let r:u32 = thread_rng().gen_range(0..10000);// Get a random number 0 <= r < 10k
         
     let mut a = [0u32; 10000];    // Array of 10k elements initialized to 0
-    for i in 0..10000 {                  // 10k outer loop iterations
+    for el in a.iter_mut() {                  // 10k outer loop iterations
         for j in 0..100000 {               // 100k inner loop iterations, per outer loop iteration
-            a[i] = a[i] + j % n;                // Simple sum
+            *el += *el + j % n;                // Simple sum
         }
-        a[i] += r;                              // Add a random value to each element in array
+        *el += r;                              // Add a random value to each element in array
     }
     println!("{}", a[r as usize]);              // Print out a single element from the array
 }