<!-- Thanks for helping to improve this project! 🙏 -->

I have:

* [X] Read the project [README](../README.md), including the [benchmark descriptions](../README.md#available-benchmarks)

## Description of changes
Change the rust loop to avoid bound checking, this is an optimisation a way that the rust officiel book tell us when we start to learn rust so it's not a cheat.

## Loop
Rust check every time if your index is higher than the lenght of the vector for memory safety.
But if we iterate over the vector, the rust compiler can undestund that we want to iterat over the vector only and it will not go futher than the lenght of the vector and ovoid bound checking.

## Levenshtein
- Less useless branch
- Reduce the lenght of the vector who was set with a x2 lenght and use int32 like the code instead of u64.

Loop : 2.0 sec -> 1.6 sec
levenshtein : 1.6 sec-> 1.0 sec