@PEZ 2025-01-10
Yay!

It looks good to me at a glance, but my Rust is not fluent at all. Please provide `run.sh` results for before and after the change, so that we easier can assess the changes.