@Paulo-21 2025-01-10
## Loop
Before : 
![image](https://github.com/user-attachments/assets/6fd594c2-8176-47cc-9bb6-649bf0571a4f)
Well i know it s wird, i m  not able to reproduce the previous time but it 's unexpected, maybe i was too tired when testing XD 
"After" :
![image](https://github.com/user-attachments/assets/5878256d-2cc0-48b1-ba43-33a7079359ef)
Just for flex  : 
![image](https://github.com/user-attachments/assets/a544728d-75e9-4118-9ec9-ba59b6d615ae)
RUSTFLAGS='-C target-cpu=native' cargo build -r 
I saw g++ using native option so why do Rust doesn't use it ?
![image](https://github.com/user-attachments/assets/8273649a-b692-41a7-bd85-451d2966fabf)

## Levenshtein
Before : 
![image](https://github.com/user-attachments/assets/67f07f4a-19a2-4ec8-a115-33cf623f5ad7)
After : 
![image](https://github.com/user-attachments/assets/8c4d039f-e365-4078-9f0a-c75a18abeedf)