@zierf 2025-01-13
In [Commit #ba17989](https://github.com/bddicken/languages/pull/335/commits/ba1798905a1c77c5e0be1f0af3f159a522f2aee0) you replaced following code:
```Rust
                let distance = levenshtein_distance(&args[i], &args[j]);

                if let Some(current_min) = min_distance {
                    if distance < current_min {
                        min_distance = Some(distance);
                    }
                } else {
                    min_distance = Some(distance);
                }
```

Interestingly, the new variant with the constant minimum allocation runs _~11%_ slower.
```Rust
                let distance = levenshtein_distance(&args[i], &args[j]);
                min_distance = std::cmp::min(distance, min_distance);
```